@@ -22,61 +22,81 @@
     "react-scripts-ts-antd": "./bin/react-scripts-ts-antd.js"
   },
   "dependencies": {
-    "autoprefixer": "7.1.6",
-    "babel-jest": "^22.1.0",
-    "babel-loader": "^7.1.2",
+    "@babel/core": "7.1.0",
+    "@svgr/webpack": "2.4.1",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-eslint": "9.0.0",
+    "babel-jest": "23.6.0",
+    "babel-loader": "8.0.4",
     "babel-plugin-import": "^1.7.0",
-    "babel-preset-react-app": "^3.1.1",
-    "case-sensitive-paths-webpack-plugin": "2.1.1",
-    "chalk": "1.1.3",
-    "css-loader": "0.28.7",
-    "dotenv": "4.0.0",
+    "babel-plugin-named-asset-import": "^0.2.0",
+    "babel-preset-react-app": "^5.0.0",
+    "bfj": "6.1.1",
+    "case-sensitive-paths-webpack-plugin": "2.1.2",
+    "chalk": "2.4.1",
+    "css-loader": "1.0.0",
+    "dotenv": "6.0.0",
     "dotenv-expand": "4.2.0",
-    "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "0.11.2",
-    "fork-ts-checker-webpack-plugin": "^0.2.8",
-    "fs-extra": "3.0.1",
-    "html-webpack-plugin": "2.29.0",
-    "jest": "22.1.4",
+    "eslint": "5.6.0",
+    "eslint-config-react-app": "^3.0.0",
+    "eslint-loader": "2.1.1",
+    "eslint-plugin-flowtype": "2.50.1",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-jsx-a11y": "6.1.1",
+    "eslint-plugin-react": "7.11.1",
+    "file-loader": "2.0.0",
+    "fork-ts-checker-webpack-plugin": "0.4.9",
+    "fs-extra": "7.0.0",
+    "html-webpack-plugin": "4.0.0-alpha.2",
+    "identity-obj-proxy": "3.0.0",
+    "jest": "23.6.0",
     "less": "^2.7.3",
     "less-loader": "^4.0.5",
-    "object-assign": "4.1.1",
-    "postcss-flexbugs-fixes": "3.2.0",
-    "postcss-loader": "2.0.8",
-    "postcss-scss": "^1.0.5",
-    "precss": "^3.1.0",
-    "promise": "8.0.1",
-    "raf": "3.4.0",
+    "mini-css-extract-plugin": "0.4.3",
+    "node-sass": "^4.9.3",
+    "optimize-css-assets-webpack-plugin": "5.0.1",
+    "postcss-flexbugs-fixes": "4.1.0",
+    "postcss-loader": "3.0.0",
+    "postcss-preset-env": "6.0.6",
+    "postcss-safe-parser": "4.0.1",
+    "react-app-polyfill": "^0.1.1",
     "react-app-rewired": "^1.4.1",
-    "react-dev-utils": "^5.0.1",
-    "resolve": "1.6.0",
+    "react-dev-utils": "^6.0.1",
+    "resolve": "1.8.1",
+    "sass-loader": "7.1.0",
     "source-map-loader": "^0.2.1",
-    "style-loader": "0.19.0",
-    "sw-precache-webpack-plugin": "0.11.4",
-    "ts-classname-plugin": "^1.0.4",
+    "style-loader": "0.23.0",
+    "terser-webpack-plugin": "1.1.0",
+    "thread-loader": "1.2.0",
+    "ts-classname-plugin": "^2.0.4",
     "ts-import-plugin": "^1.4.3",
     "ts-jest": "22.0.1",
-    "ts-loader": "^2.3.7",
+    "ts-loader": "4.x.x",
     "tsconfig-paths-webpack-plugin": "^2.0.0",
     "tslint": "^5.7.0",
     "tslint-config-prettier": "^1.10.0",
     "tslint-react": "^3.2.0",
-    "uglifyjs-webpack-plugin": "^1.1.8",
-    "url-loader": "0.6.2",
-    "webpack": "3.8.1",
-    "webpack-dev-server": "2.9.4",
-    "webpack-manifest-plugin": "1.3.2",
-    "whatwg-fetch": "2.0.3"
+    "url-loader": "1.1.1",
+    "webpack": "4.19.1",
+    "webpack-dev-server": "3.1.9",
+    "webpack-manifest-plugin": "2.0.4",
+    "workbox-webpack-plugin": "3.6.1"
   },
   "devDependencies": {
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "typescript": "^2.7.1"
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
+    "typescript": "3.0.1"
   },
   "peerDependencies": {
-    "typescript": "2.x"
+    "typescript": "3.x.x"
   },
   "optionalDependencies": {
-    "fsevents": "^1.1.3"
-  }
+    "fsevents": "1.2.4"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
